debounced throttle implementation + demo
This is an implementation of a debounce + throttle strategy
for the port communication. This could be implemented as a store enhancer
and the same strategy could be applied to all store changes. This could
also optimized for store savings (if we were to save the state on state change)

The idea is, that we want to debounce the propagation of our state changes
if the changes comes in <100ms intervals. But if a lot of changes are coming
then we can end up waiting for the state to settle for 5+ seconds. Thus the
need for a throttle function, so we can be sure that the app stays responsive.

I have chosen 100ms for debounceDuration and 500ms for the periodic communication.

This implementation is a demo for a conversation about this.
Not meant to be production ready.

The port communication can fill up quickly and can result in
slow communication between UI and background thus the
app can become unresponsive. Both the speed of the messages and
the size of them are important. But it seems that the speed in which the
communication comes has bigger effect on the throughput.